backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.40.0
PyPDF2==3.0.1
python-docx==1.1.0
werkzeug==3.0.1
stripe==7.9.0
python-dotenv==1.0.0
